ecoDashV2-main/requirements.txt
gunicorn==20.1.0
dash==2.7.1
pandas==2.0.2
#google==3.0.0
cloud-sql-python-connector==1.4.3
sqlalchemy==2.0.23
supabase==1.0.4
plotly==5.11.0
layout==0.4.1
pymysql==1.1.0
psycopg2-binary==2.9.9
google-cloud-secret-manager ==	2.17.0

